@@ -1,20 +1,20 @@
 {
   "name": "mysql2",
-  "version": "2.3.3",
+  "version": "3.9.1",
   "description": "fast mysql driver. Implements core protocol, prepared statements, ssl and compression in native JS",
   "main": "index.js",
-  "directories": {
-    "example": "examples"
-  },
+  "typings": "typings/mysql/index",
   "scripts": {
     "lint": "npm run lint:docs && npm run lint:code",
-    "lint:code": "eslint index.js promise.js \"lib/**/*.js\" \"test/**/*.js\" \"benchmarks/**/*.js\"",
-    "lint:docs": "eslint Contributing.md \"documentation/**/*.md\" \"examples/*.js\"",
+    "lint:code": "eslint index.js promise.js index.d.ts promise.d.ts \"typings/**/*.ts\" \"lib/**/*.js\" \"test/**/*.{js,ts}\" \"benchmarks/**/*.js\"",
+    "lint:docs": "eslint Contributing.md README.md",
     "test": "node ./test/run.js",
+    "test:builtin-node-runner": "NODE_V8_COVERAGE=./coverage node --test --experimental-test-coverage test/builtin-runner",
+    "test:tsc-build": "cd \"test/tsc-build\" && npx tsc -p \"tsconfig.json\"",
     "coverage-test": "c8 -r cobertura -r lcov -r text node ./test/run.js",
     "benchmark": "node ./benchmarks/benchmark.js",
-    "prettier": "prettier --single-quote --trailing-comma none --write \"{lib,examples,test}/**/*.js\"",
-    "prettier:docs": "prettier --single-quote --trailing-comma none --write README.md documentation/*",
+    "prettier": "prettier --single-quote --trailing-comma none --write \"{lib,test}/**/*.js\"",
+    "prettier:docs": "prettier --single-quote --trailing-comma none --write README.md",
     "precommit": "lint-staged",
     "eslint-check": "eslint --print-config .eslintrc | eslint-config-prettier-check",
     "wait-port": "wait-on"
@@ -29,6 +29,7 @@
     "type": "git",
     "url": "https://github.com/sidorares/node-mysql2"
   },
+  "homepage": "https://sidorares.github.io/node-mysql2/docs",
   "keywords": [
     "mysql",
     "client",
@@ -36,7 +37,7 @@
   ],
   "files": [
     "lib",
-    "typings",
+    "typings/mysql",
     "index.js",
     "index.d.ts",
     "promise.js",
@@ -44,42 +45,43 @@
   ],
   "exports": {
     ".": "./index.js",
+    "./package.json": "./package.json",
     "./promise": "./promise.js",
     "./promise.js": "./promise.js"
   },
   "engines": {
     "node": ">= 8.0"
   },
-  "author": "Andrey Sidorov <sidorares@yandex.ru>",
+  "author": "Andrey Sidorov <andrey.sidorov@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "denque": "^2.0.1",
+    "denque": "^2.1.0",
     "generate-function": "^2.3.1",
     "iconv-lite": "^0.6.3",
-    "long": "^4.0.0",
-    "lru-cache": "^6.0.0",
-    "named-placeholders": "^1.1.2",
+    "long": "^5.2.1",
+    "lru-cache": "^8.0.0",
+    "named-placeholders": "^1.1.3",
     "seq-queue": "^0.0.5",
     "sqlstring": "^2.3.2"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
+    "@types/node": "^20.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.42.1",
+    "@typescript-eslint/parser": "^5.42.1",
     "assert-diff": "^3.0.2",
     "benchmark": "^2.1.4",
-    "c8": "^7.10.0",
+    "c8": "^9.0.0",
     "error-stack-parser": "^2.0.3",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint": "^8.27.0",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-async-await": "0.0.0",
-    "eslint-plugin-markdown": "^2.2.1",
-    "husky": "^7.0.2",
-    "is-async-supported": "^1.2.0",
-    "lint-staged": "^11.2.0",
+    "eslint-plugin-markdown": "^3.0.0",
+    "husky": "^9.0.2",
+    "lint-staged": "^15.0.1",
     "portfinder": "^1.0.28",
-    "prettier": "^2.4.1",
+    "prettier": "^3.0.0",
     "progress": "^2.0.3",
-    "typescript": "^4.4.3",
+    "typescript": "^5.0.2",
     "urun": "0.0.8",
     "utest": "0.0.8"
   }
