@@ -2,7 +2,7 @@
 
 const PLUGIN_NAME = 'sha256_password';
 const crypto = require('crypto');
-const { xor } = require('../auth_41');
+const { xorRotating } = require('../auth_41');
 
 const REQUEST_SERVER_KEY_PACKET = Buffer.from([1]);
 
@@ -11,9 +11,9 @@ const STATE_WAIT_SERVER_KEY = 1;
 const STATE_FINAL = -1;
 
 function encrypt(password, scramble, key) {
-  const stage1 = xor(
-    Buffer.from(`${password}\0`, 'utf8').toString('binary'),
-    scramble.toString('binary')
+  const stage1 = xorRotating(
+    Buffer.from(`${password}\0`, 'utf8'),
+    scramble
   );
   return crypto.publicEncrypt(key, stage1);
 }
