@@ -9,11 +9,13 @@ const Packets = require('../packets/index.js');
 const sha256_password = require('../auth_plugins/sha256_password');
 const caching_sha2_password = require('../auth_plugins/caching_sha2_password.js');
 const mysql_native_password = require('../auth_plugins/mysql_native_password.js');
+const mysql_clear_password = require('../auth_plugins/mysql_clear_password.js');
 
 const standardAuthPlugins = {
   sha256_password: sha256_password({}),
   caching_sha2_password: caching_sha2_password({}),
-  mysql_native_password: mysql_native_password({})
+  mysql_native_password: mysql_native_password({}),
+  mysql_clear_password: mysql_clear_password({})
 };
 
 function warnLegacyAuthSwitch() {
