@@ -3,6 +3,11 @@
 // emitted in the command instance itself.
 // Modifications copyright (c) 2021, Oracle and/or its affiliates.
 
+// This file was modified by Oracle on September 21, 2021.
+// Handshake workflow now supports additional authentication factors requested
+// by the server.
+// Modifications copyright (c) 2021, Oracle and/or its affiliates.
+
 'use strict';
 
 const Command = require('./command.js');
@@ -26,6 +31,7 @@ class ClientHandshake extends Command {
     super();
     this.handshake = null;
     this.clientFlags = clientFlags;
+    this.authenticationFactor = 0;
   }
 
   start() {
@@ -51,6 +57,14 @@ class ClientHandshake extends Command {
     }
     this.user = connection.config.user;
     this.password = connection.config.password;
+    // "password1" is an alias to the original "password" value
+    // to make it easier to integrate multi-factor authentication
+    this.password1 = connection.config.password;
+    // "password2" and "password3" are the 2nd and 3rd factor authentication
+    // passwords, which can be undefined depending on the authentication
+    // plugin being used
+    this.password2 = connection.config.password2;
+    this.password3 = connection.config.password3;
     this.passwordSha1 = connection.config.passwordSha1;
     this.database = connection.config.database;
     this.autPluginName = this.handshake.autPluginName;
@@ -109,6 +123,12 @@ class ClientHandshake extends Command {
     connection.connectionId = this.handshake.connectionId;
     const serverSSLSupport =
       this.handshake.capabilityFlags & ClientConstants.SSL;
+    // multi factor authentication is enabled with the
+    // "MULTI_FACTOR_AUTHENTICATION" capability and should only be used if it
+    // is supported by the server
+    const multiFactorAuthentication =
+      this.handshake.capabilityFlags & ClientConstants.MULTI_FACTOR_AUTHENTICATION;
+    this.clientFlags = this.clientFlags | multiFactorAuthentication;
     // use compression only if requested by client and supported by server
     connection.config.compress =
       connection.config.compress &&
@@ -117,7 +137,7 @@ class ClientHandshake extends Command {
     if (connection.config.ssl) {
       // client requires SSL but server does not support it
       if (!serverSSLSupport) {
-        const err = new Error('Server does not support secure connnection');
+        const err = new Error('Server does not support secure connection');
         err.code = 'HANDSHAKE_NO_SSL_SUPPORT';
         err.fatal = true;
         this.emit('error', err);
@@ -141,17 +161,36 @@ class ClientHandshake extends Command {
     } else {
       this.sendCredentials(connection);
     }
+    if (multiFactorAuthentication) {
+      // if the server supports multi-factor authentication, we enable it in
+      // the client
+      this.authenticationFactor = 1;
+    }
     return ClientHandshake.prototype.handshakeResult;
   }
 
   handshakeResult(packet, connection) {
     const marker = packet.peekByte();
-    if (marker === 0xfe || marker === 1) {
+    // packet can be OK_Packet, ERR_Packet, AuthSwitchRequest, AuthNextFactor
+    // or AuthMoreData
+    if (marker === 0xfe || marker === 1 || marker === 0x02) {
       const authSwitch = require('./auth_switch');
       try {
         if (marker === 1) {
           authSwitch.authSwitchRequestMoreData(packet, connection, this);
         } else {
+          // if authenticationFactor === 0, it means the server does not support
+          // the multi-factor authentication capability
+          if (this.authenticationFactor !== 0) {
+            // if we are past the first authentication factor, we should use the
+            // corresponding password (if there is one)
+            connection.config.password = this[`password${this.authenticationFactor}`];
+            // update the current authentication factor
+            this.authenticationFactor += 1;
+          }
+          // if marker === 0x02, it means it is an AuthNextFactor packet,
+          // which is similar in structure to an AuthSwitchRequest packet,
+          // so, we can use it directly
           authSwitch.authSwitchRequest(packet, connection, this);
         }
         return ClientHandshake.prototype.handshakeResult;
