@@ -7,7 +7,7 @@ const Packet = require('../packets/packet.js');
 class Quit extends Command {
   constructor(callback) {
     super();
-    this.done = callback;
+    this.onResult = callback;
   }
 
   start(packet, connection) {
@@ -18,8 +18,8 @@ class Quit extends Command {
       0,
       5
     );
-    if (this.done) {
-      this.done();
+    if (this.onResult) {
+      this.onResult();
     }
     connection.writePacket(quit);
     return null;
