@@ -69,12 +69,41 @@ class ServerHandshake extends Command {
     return ServerHandshake.prototype.dispatchCommands;
   }
 
+  _isStatement(query, name) {
+    const firstWord = query.split(' ')[0].toUpperCase();
+    return firstWord === name;
+  }
+
   dispatchCommands(packet, connection) {
     // command from client to server
     let knownCommand = true;
     const encoding = connection.clientHelloReply.encoding;
     const commandCode = packet.readInt8();
     switch (commandCode) {
+      case CommandCode.STMT_PREPARE:
+        if (connection.listeners('stmt_prepare').length) {
+          const query = packet.readString(undefined, encoding);
+          connection.emit('stmt_prepare', query);
+        } else {
+          connection.writeError({
+            code: Errors.HA_ERR_INTERNAL_ERROR,
+            message:
+              'No query handler for prepared statements.'
+          });
+        }
+        break;
+      case CommandCode.STMT_EXECUTE:
+        if (connection.listeners('stmt_execute').length) {
+          const { stmtId, flags, iterationCount, values } = Packets.Execute.fromPacket(packet, encoding);
+          connection.emit('stmt_execute', stmtId, flags, iterationCount, values);
+        } else {
+          connection.writeError({
+            code: Errors.HA_ERR_INTERNAL_ERROR,
+            message:
+              'No query handler for execute statements.'
+          });
+        }
+        break;
       case CommandCode.QUIT:
         if (connection.listeners('quit').length) {
           connection.emit('quit');
@@ -93,7 +122,13 @@ class ServerHandshake extends Command {
       case CommandCode.QUERY:
         if (connection.listeners('query').length) {
           const query = packet.readString(undefined, encoding);
-          connection.emit('query', query);
+          if (this._isStatement(query, 'PREPARE') || this._isStatement(query, 'SET')) {
+            connection.emit('stmt_prepare', query);
+          }
+          else if (this._isStatement(query, 'EXECUTE')) {
+            connection.emit('stmt_execute', null, null, null, null, query);
+          }
+          else connection.emit('query', query);
         } else {
           connection.writeError({
             code: Errors.HA_ERR_INTERNAL_ERROR,
@@ -103,7 +138,7 @@ class ServerHandshake extends Command {
         break;
       case CommandCode.FIELD_LIST:
         if (connection.listeners('field_list').length) {
-          const table = packet.readNullTerminatedString();
+          const table = packet.readNullTerminatedString(encoding);
           const fields = packet.readString(undefined, encoding);
           connection.emit('field_list', table, fields);
         } else {
