@@ -8,6 +8,11 @@
 // there is a fatal error.
 // Modifications copyright (c) 2021, Oracle and/or its affiliates.
 
+// This file was modified by Oracle on September 21, 2021.
+// The changes involve passing additional authentication factor passwords
+// to the ChangeUser Command instance.
+// Modifications copyright (c) 2021, Oracle and/or its affiliates.
+
 'use strict';
 
 const Net = require('net');
@@ -17,7 +22,7 @@ const EventEmitter = require('events').EventEmitter;
 const Readable = require('stream').Readable;
 const Queue = require('denque');
 const SqlString = require('sqlstring');
-const LRU = require('lru-cache');
+const LRU = require('lru-cache').default;
 
 const PacketParser = require('./packet_parser.js');
 const Packets = require('./packets/index.js');
@@ -47,9 +52,16 @@ class Connection extends EventEmitter {
           opts.config.host
         );
 
-        // Enable keep-alive on the socket.  It's disabled by default, but the
-        // user can enable it and supply an initial delay.
-        this.stream.setKeepAlive(true, this.config.keepAliveInitialDelay);
+        // Optionally enable keep-alive on the socket.
+        if (this.config.enableKeepAlive) {
+          this.stream.on('connect', () => {
+            this.stream.setKeepAlive(true, this.config.keepAliveInitialDelay);
+          });
+        }
+
+        // Enable TCP_NODELAY flag. This is needed so that the network packets
+        // are sent immediately to the server
+        this.stream.setNoDelay(true);
       }
       // if stream is a function, treat it as "stream agent / factory"
     } else if (typeof opts.config.stream === 'function')  {
@@ -65,7 +77,7 @@ class Connection extends EventEmitter {
     this._paused_packets = new Queue();
     this._statements = new LRU({
       max: this.config.maxPreparedStatements,
-      dispose: function(key, statement) {
+      dispose: function(statement) {
         statement.close();
       }
     });
@@ -91,6 +103,10 @@ class Connection extends EventEmitter {
       }
       this.packetParser.execute(data);
     });
+    this.stream.on('end', () => {
+      // emit the end event so that the pooled connection can close the connection
+      this.emit('end');
+    });
     this.stream.on('close', () => {
       // we need to set this flag everywhere where we want connection to close
       if (this._closing) {
@@ -112,7 +128,7 @@ class Connection extends EventEmitter {
       handshakeCommand.on('end', () => {
         // this happens when handshake finishes early either because there was
         // some fatal error or the server sent an error packet instead of
-        // an hello packet (for example, 'Too many connactions' error)
+        // an hello packet (for example, 'Too many connections' error)
         if (!handshakeCommand.handshake || this._fatalError || this._protocolError) {
           return;
         }
@@ -126,9 +142,9 @@ class Connection extends EventEmitter {
       });
       this.addCommand(handshakeCommand);
     }
-    // in case there was no initiall handshake but we need to read sting, assume it utf-8
+    // in case there was no initial handshake but we need to read sting, assume it utf-8
     // most common example: "Too many connections" error ( packet is sent immediately on connection attempt, we don't know server encoding yet)
-    // will be overwrittedn with actial encoding value as soon as server handshake packet is received
+    // will be overwritten with actual encoding value as soon as server handshake packet is received
     this.serverEncoding = 'utf8';
     if (this.config.connectTimeout) {
       const timeoutHandler = this._handleTimeoutError.bind(this);
@@ -334,52 +350,51 @@ class Connection extends EventEmitter {
       ciphers: this.config.ssl.ciphers,
       key: this.config.ssl.key,
       passphrase: this.config.ssl.passphrase,
-      minVersion: this.config.ssl.minVersion
+      minVersion: this.config.ssl.minVersion,
+      maxVersion: this.config.ssl.maxVersion
     });
     const rejectUnauthorized = this.config.ssl.rejectUnauthorized;
+    const verifyIdentity = this.config.ssl.verifyIdentity;
+    const servername = this.config.host;
+
     let secureEstablished = false;
-    const secureSocket = new Tls.TLSSocket(this.stream, {
-      rejectUnauthorized: rejectUnauthorized,
-      requestCert: true,
-      secureContext: secureContext,
-      isServer: false
+    this.stream.removeAllListeners('data');
+    const secureSocket = Tls.connect({
+      rejectUnauthorized,
+      requestCert: rejectUnauthorized,
+      checkServerIdentity: verifyIdentity
+        ? Tls.checkServerIdentity
+        : function() { return undefined; },
+      secureContext,
+      isServer: false,
+      socket: this.stream,
+      servername
+    }, () => {
+      secureEstablished = true;
+      if (rejectUnauthorized) {
+        if (typeof servername === 'string' && verifyIdentity) {
+          const cert = secureSocket.getPeerCertificate(true);
+          const serverIdentityCheckError = Tls.checkServerIdentity(servername, cert);
+          if (serverIdentityCheckError) {
+            onSecure(serverIdentityCheckError);
+            return;
+          }
+        }
+      }
+      onSecure();
     });
     // error handler for secure socket
-    secureSocket.on('_tlsError', err => {
+    secureSocket.on('error', err => {
       if (secureEstablished) {
         this._handleNetworkError(err);
       } else {
         onSecure(err);
       }
     });
-    secureSocket.on('secure', () => {
-      secureEstablished = true;
-      onSecure(rejectUnauthorized ? secureSocket.ssl.verifyError() : null);
-    });
     secureSocket.on('data', data => {
       this.packetParser.execute(data);
     });
-    this.write = buffer => {
-      secureSocket.write(buffer);
-    };
-    // start TLS communications
-    secureSocket._start();
-  }
-
-  pipe() {
-    if (this.stream instanceof Net.Stream) {
-      this.stream.ondata = (data, start, end) => {
-        this.packetParser.execute(data, start, end);
-      };
-    } else {
-      this.stream.on('data', data => {
-        this.packetParser.execute(
-          data.parent,
-          data.offset,
-          data.offset + data.length
-        );
-      });
-    }
+    this.write = buffer => secureSocket.write(buffer);
   }
 
   protocolError(message, code) {
@@ -396,24 +411,15 @@ class Connection extends EventEmitter {
     err.code = code || 'PROTOCOL_ERROR';
     this.emit('error', err);
   }
+  
+  get fatalError() {
+    return this._fatalError;
+  }
 
   handlePacket(packet) {
     if (this._paused) {
       this._paused_packets.push(packet);
       return;
-    }
-    if (packet) {
-      if (this.sequenceId !== packet.sequenceId) {
-        const err = new Error(
-          `Warning: got packets out of order. Expected ${this.sequenceId} but received ${packet.sequenceId}`
-        );
-        err.expected = this.sequenceId;
-        err.received = packet.sequenceId;
-        this.emit('warn', err); // REVIEW
-        // eslint-disable-next-line no-console
-        console.error(err.message);
-      }
-      this._bumpSequenceId(packet.numPackets);
     }
     if (this.config.debug) {
       if (packet) {
@@ -453,14 +459,37 @@ class Connection extends EventEmitter {
       this.close();
       return;
     }
-    const done = this._command.execute(packet, this);
-    if (done) {
-      this._command = this._commands.shift();
-      if (this._command) {
-        this.sequenceId = 0;
-        this.compressedSequenceId = 0;
-        this.handlePacket();
+    if (packet) {
+      // Note: when server closes connection due to inactivity, Err packet ER_CLIENT_INTERACTION_TIMEOUT from MySQL 8.0.24, sequenceId will be 0
+      if (this.sequenceId !== packet.sequenceId) {
+        const err = new Error(
+          `Warning: got packets out of order. Expected ${this.sequenceId} but received ${packet.sequenceId}`
+        );
+        err.expected = this.sequenceId;
+        err.received = packet.sequenceId;
+        this.emit('warn', err); // REVIEW
+        // eslint-disable-next-line no-console
+        console.error(err.message);
       }
+      this._bumpSequenceId(packet.numPackets);
+    }
+    try {
+      if (this._fatalError) {
+        // skip remaining packets after client is in the error state
+        return;
+      }
+      const done = this._command.execute(packet, this);
+      if (done) {
+        this._command = this._commands.shift();
+        if (this._command) {
+          this.sequenceId = 0;
+          this.compressedSequenceId = 0;
+          this.handlePacket();
+        }
+      }
+    } catch (err) {
+      this._handleFatalError(err);
+      this.stream.destroy();
     }
   }
 
@@ -556,7 +585,7 @@ class Connection extends EventEmitter {
     this._paused = false;
     while ((packet = this._paused_packets.shift())) {
       this.handlePacket(packet);
-      // don't resume if packet hander paused connection
+      // don't resume if packet handler paused connection
       if (this._paused) {
         return;
       }
@@ -582,17 +611,24 @@ class Connection extends EventEmitter {
     const key = Connection.statementKey(options);
     const stmt = this._statements.get(key);
     if (stmt) {
-      this._statements.del(key);
+      this._statements.delete(key);
       stmt.close();
     }
     return stmt;
   }
 
   execute(sql, values, cb) {
-    let options = {};
+    let options = {
+      infileStreamFactory: this.config.infileStreamFactory
+    };
     if (typeof sql === 'object') {
       // execute(options, cb)
-      options = sql;
+      options = {
+        ...options,
+        ...sql,
+        sql: sql.sql,
+        values: sql.values
+      };
       if (typeof values === 'function') {
         cb = values;
       } else {
@@ -671,7 +707,12 @@ class Connection extends EventEmitter {
       new Commands.ChangeUser(
         {
           user: options.user || this.config.user,
-          password: options.password || this.config.password,
+          // for the purpose of multi-factor authentication, or not, the main
+          // password (used for the 1st authentication factor) can also be
+          // provided via the "password1" option
+          password: options.password || options.password1 || this.config.password || this.config.password1,
+          password2: options.password2 || this.config.password2,
+          password3: options.password3 || this.config.password3,
           passwordSha1: options.passwordSha1 || this.config.passwordSha1,
           database: options.database || this.config.database,
           timeout: options.timeout,
@@ -804,14 +845,23 @@ class Connection extends EventEmitter {
     );
   }
 
-  writeTextResult(rows, columns) {
+  writeBinaryRow(column) {
+    this.writePacket(
+      Packets.BinaryRow.toPacket(column, this.serverConfig.encoding)
+    );
+  }
+
+  writeTextResult(rows, columns, binary=false) {
     this.writeColumns(columns);
     rows.forEach(row => {
       const arrayRow = new Array(columns.length);
       columns.forEach(column => {
         arrayRow.push(row[column.name]);
       });
-      this.writeTextRow(arrayRow);
+      if(binary) {
+        this.writeBinaryRow(arrayRow);
+      }
+      else this.writeTextRow(arrayRow);
     });
     this.writeEof();
   }
@@ -859,11 +909,17 @@ class Connection extends EventEmitter {
 
   static createQuery(sql, values, cb, config) {
     let options = {
-      rowsAsArray: config.rowsAsArray
+      rowsAsArray: config.rowsAsArray,
+      infileStreamFactory: config.infileStreamFactory
     };
     if (typeof sql === 'object') {
       // query(options, cb)
-      options = sql;
+      options = {
+        ...options,
+        ...sql,
+        sql: sql.sql,
+        values: sql.values
+      };
       if (typeof values === 'function') {
         cb = values;
       } else if (values !== undefined) {
@@ -887,50 +943,6 @@ class Connection extends EventEmitter {
       `${typeof options.nestTables}/${options.nestTables}/${options.rowsAsArray}${options.sql}`
     );
   }
-}
-
-if (Tls.TLSSocket) {
-  // not supported
-} else {
-  Connection.prototype.startTLS = function _startTLS(onSecure) {
-    if (this.config.debug) {
-      // eslint-disable-next-line no-console
-      console.log('Upgrading connection to TLS');
-    }
-    const crypto = require('crypto');
-    const config = this.config;
-    const stream = this.stream;
-    const rejectUnauthorized = this.config.ssl.rejectUnauthorized;
-    const credentials = crypto.createCredentials({
-      key: config.ssl.key,
-      cert: config.ssl.cert,
-      passphrase: config.ssl.passphrase,
-      ca: config.ssl.ca,
-      ciphers: config.ssl.ciphers
-    });
-    const securePair = Tls.createSecurePair(
-      credentials,
-      false,
-      true,
-      rejectUnauthorized
-    );
-
-    if (stream.ondata) {
-      stream.ondata = null;
-    }
-    stream.removeAllListeners('data');
-    stream.pipe(securePair.encrypted);
-    securePair.encrypted.pipe(stream);
-    securePair.cleartext.on('data', data => {
-      this.packetParser.execute(data);
-    });
-    this.write = function(buffer) {
-      securePair.cleartext.write(buffer);
-    };
-    securePair.on('secure', () => {
-      onSecure(rejectUnauthorized ? securePair.ssl.verifyError() : null);
-    });
-  };
 }
 
 module.exports = Connection;
