@@ -1,8 +1,16 @@
+// This file was modified by Oracle on September 21, 2021.
+// New connection options for additional authentication factors were
+// introduced.
+// Multi-factor authentication capability is now enabled if one of these
+// options is used.
+// Modifications copyright (c) 2021, Oracle and/or its affiliates.
+
 'use strict';
 
 const { URL } = require('url');
 const ClientConstants = require('./constants/client');
 const Charsets = require('./constants/charsets');
+const { version } = require('../package.json')
 let SSLProfiles = null;
 
 const validOptions = {
@@ -22,6 +30,7 @@ const validOptions = {
   flags: 1,
   host: 1,
   insecureAuth: 1,
+  infileStreamFactory: 1,
   isServer: 1,
   keepAliveInitialDelay: 1,
   localAddress: 1,
@@ -30,6 +39,11 @@ const validOptions = {
   namedPlaceholders: 1,
   nestTables: 1,
   password: 1,
+  // with multi-factor authentication, the main password (used for the first
+  // authentication factor) can be provided via password1
+  password1: 1,
+  password2: 1,
+  password3: 1,
   passwordSha1: 1,
   pool: 1,
   port: 1,
@@ -48,6 +62,8 @@ const validOptions = {
   user: 1,
   // These options are used for Pool
   connectionLimit: 1,
+  maxIdle: 1,
+  idleTimeout: 1,
   Promise: 1,
   queueLimit: 1,
   waitForConnections: 1
@@ -78,17 +94,23 @@ class ConnectionConfig {
     this.isServer = options.isServer;
     this.stream = options.stream;
     this.host = options.host || 'localhost';
-    this.port = options.port || 3306;
+    this.port = (typeof options.port === 'string' ? parseInt(options.port, 10) : options.port)|| 3306;
     this.localAddress = options.localAddress;
     this.socketPath = options.socketPath;
     this.user = options.user || undefined;
-    this.password = options.password || undefined;
+    // for the purpose of multi-factor authentication, or not, the main
+    // password (used for the 1st authentication factor) can also be
+    // provided via the "password1" option
+    this.password = options.password || options.password1 || undefined;
+    this.password2 = options.password2 || undefined;
+    this.password3 = options.password3 || undefined;
     this.passwordSha1 = options.passwordSha1 || undefined;
     this.database = options.database;
     this.connectTimeout = isNaN(options.connectTimeout)
       ? 10 * 1000
       : options.connectTimeout;
     this.insecureAuth = options.insecureAuth || false;
+    this.infileStreamFactory = options.infileStreamFactory || undefined;
     this.supportBigNumbers = options.supportBigNumbers || false;
     this.bigNumberStrings = options.bigNumberStrings || false;
     this.decimalNumbers = options.decimalNumbers || false;
@@ -96,7 +118,7 @@ class ConnectionConfig {
     this.debug = options.debug;
     this.trace = options.trace !== false;
     this.stringifyObjects = options.stringifyObjects || false;
-    this.enableKeepAlive = !!options.enableKeepAlive;
+    this.enableKeepAlive = options.enableKeepAlive !== false;
     this.keepAliveInitialDelay = options.keepAliveInitialDelay || 0;
     if (
       options.timezone &&
@@ -130,7 +152,7 @@ class ConnectionConfig {
       // "+" is a url encoded char for space so it
       // gets translated to space when giving a
       // connection string..
-      this.timezone = `+${this.timezone.substr(1)}`;
+      this.timezone = `+${this.timezone.slice(1)}`;
     }
     if (this.ssl) {
       if (typeof this.ssl !== 'object') {
@@ -152,7 +174,13 @@ class ConnectionConfig {
       ConnectionConfig.getDefaultFlags(options),
       options.flags || ''
     );
-    this.connectAttributes = options.connectAttributes;
+    // Default connection attributes
+    // https://dev.mysql.com/doc/refman/8.0/en/performance-schema-connection-attribute-tables.html
+    const defaultConnectAttributes =  {
+      _client_name: 'Node-MySQL-2',
+      _client_version: version
+    };
+    this.connectAttributes = { ...defaultConnectAttributes, ...(options.connectAttributes || {})};
     this.maxPreparedStatements = options.maxPreparedStatements || 16000;
   }
 
@@ -200,7 +228,8 @@ class ConnectionConfig {
       'SECURE_CONNECTION',
       'MULTI_RESULTS',
       'TRANSACTIONS',
-      'SESSION_TRACK'
+      'SESSION_TRACK',
+      'CONNECT_ATTRS'
     ];
     if (options && options.multipleStatements) {
       defaultFlags.push('MULTI_STATEMENTS');
@@ -208,9 +237,6 @@ class ConnectionConfig {
     defaultFlags.push('PLUGIN_AUTH');
     defaultFlags.push('PLUGIN_AUTH_LENENC_CLIENT_DATA');
 
-    if (options && options.connectAttributes) {
-      defaultFlags.push('CONNECT_ATTRS');
-    }
     return defaultFlags;
   }
 
@@ -236,11 +262,11 @@ class ConnectionConfig {
   static parseUrl(url) {
     const parsedUrl = new URL(url);
     const options = {
-      host: parsedUrl.hostname,
-      port: parsedUrl.port,
-      database: parsedUrl.pathname.substr(1),
-      user: parsedUrl.username,
-      password: parsedUrl.password
+      host: decodeURIComponent(parsedUrl.hostname),
+      port: parseInt(parsedUrl.port, 10),
+      database: decodeURIComponent(parsedUrl.pathname.slice(1)),
+      user: decodeURIComponent(parsedUrl.username),
+      password: decodeURIComponent(parsedUrl.password),
     };
     parsedUrl.searchParams.forEach((value, key) => {
       try {
