@@ -104,7 +104,7 @@ module.exports = [
   'eucjpms',
   'eucjpms',
   'cp1250',
-  'utf8',
+  'utf16',
   'utf16',
   'utf16',
   'utf16',
@@ -255,6 +255,9 @@ module.exports = [
   'gb18030',
   'gb18030',
   'gb18030',
+  'utf8',
+  'utf8',
+  'utf8',
   'utf8',
   'utf8',
   'utf8',
