@@ -223,28 +223,56 @@ exports.UTF8MB4_VIETNAMESE_CI = 247;
 exports.GB18030_CHINESE_CI = 248;
 exports.GB18030_BIN = 249;
 exports.GB18030_UNICODE_520_CI = 250;
-exports.UTF8_GENERAL50_CI = 253;
+exports.UTF8_GENERAL50_CI = 253; // deprecated
 exports.UTF8MB4_0900_AI_CI = 255;
-exports.UTF8MB4_CS_0900_AI_CI = 266;
-exports.UTF8MB4_DA_0900_AI_CI = 267;
 exports.UTF8MB4_DE_PB_0900_AI_CI = 256;
-exports.UTF8MB4_EO_0900_AI_CI = 273;
-exports.UTF8MB4_ES_0900_AI_CI = 263;
-exports.UTF8MB4_ES_TRAD_0900_AI_CI = 270;
-exports.UTF8MB4_ET_0900_AI_CI = 262;
-exports.UTF8MB4_HR_0900_AI_CI = 275;
-exports.UTF8MB4_HU_0900_AI_CI = 274;
 exports.UTF8MB4_IS_0900_AI_CI = 257;
-exports.UTF8MB4_LA_0900_AI_CI = 271;
-exports.UTF8MB4_LT_0900_AI_CI = 268;
 exports.UTF8MB4_LV_0900_AI_CI = 258;
-exports.UTF8MB4_PL_0900_AI_CI = 261;
 exports.UTF8MB4_RO_0900_AI_CI = 259;
-exports.UTF8MB4_SK_0900_AI_CI = 269;
 exports.UTF8MB4_SL_0900_AI_CI = 260;
+exports.UTF8MB4_PL_0900_AI_CI = 261;
+exports.UTF8MB4_ET_0900_AI_CI = 262;
+exports.UTF8MB4_ES_0900_AI_CI = 263;
 exports.UTF8MB4_SV_0900_AI_CI = 264;
 exports.UTF8MB4_TR_0900_AI_CI = 265;
+exports.UTF8MB4_CS_0900_AI_CI = 266;
+exports.UTF8MB4_DA_0900_AI_CI = 267;
+exports.UTF8MB4_LT_0900_AI_CI = 268;
+exports.UTF8MB4_SK_0900_AI_CI = 269;
+exports.UTF8MB4_ES_TRAD_0900_AI_CI = 270;
+exports.UTF8MB4_LA_0900_AI_CI = 271;
+exports.UTF8MB4_EO_0900_AI_CI = 273;
+exports.UTF8MB4_HU_0900_AI_CI = 274;
+exports.UTF8MB4_HR_0900_AI_CI = 275;
 exports.UTF8MB4_VI_0900_AI_CI = 277;
+exports.UTF8MB4_0900_AS_CS = 278;
+exports.UTF8MB4_DE_PB_0900_AS_CS = 279;
+exports.UTF8MB4_IS_0900_AS_CS = 280;
+exports.UTF8MB4_LV_0900_AS_CS = 281;
+exports.UTF8MB4_RO_0900_AS_CS = 282;
+exports.UTF8MB4_SL_0900_AS_CS = 283;
+exports.UTF8MB4_PL_0900_AS_CS = 284;
+exports.UTF8MB4_ET_0900_AS_CS = 285;
+exports.UTF8MB4_ES_0900_AS_CS = 286;
+exports.UTF8MB4_SV_0900_AS_CS = 287;
+exports.UTF8MB4_TR_0900_AS_CS = 288;
+exports.UTF8MB4_CS_0900_AS_CS = 289;
+exports.UTF8MB4_DA_0900_AS_CS = 290;
+exports.UTF8MB4_LT_0900_AS_CS = 291;
+exports.UTF8MB4_SK_0900_AS_CS = 292;
+exports.UTF8MB4_ES_TRAD_0900_AS_CS = 293;
+exports.UTF8MB4_LA_0900_AS_CS = 294;
+exports.UTF8MB4_EO_0900_AS_CS = 296;
+exports.UTF8MB4_HU_0900_AS_CS = 297;
+exports.UTF8MB4_HR_0900_AS_CS = 298;
+exports.UTF8MB4_VI_0900_AS_CS = 300;
+exports.UTF8MB4_JA_0900_AS_CS = 303;
+exports.UTF8MB4_JA_0900_AS_CS_KS = 304;
+exports.UTF8MB4_0900_AS_CI = 305;
+exports.UTF8MB4_RU_0900_AI_CI = 306;
+exports.UTF8MB4_RU_0900_AS_CS = 307;
+exports.UTF8MB4_ZH_0900_AS_CS = 308;
+exports.UTF8MB4_0900_BIN = 309;
 
 // short aliases
 exports.BIG5 = exports.BIG5_CHINESE_CI;
