@@ -25,7 +25,9 @@ exports.srcEscape = srcEscape;
 let highlightFn;
 let cardinalRecommended = false;
 try {
-  highlightFn = require('cardinal').highlight;
+  // the purpose of this is to prevent projects using Webpack from displaying a warning during runtime if cardinal is not a dependency
+  const REQUIRE_TERMINATOR = '';
+  highlightFn = require(`cardinal${REQUIRE_TERMINATOR}`).highlight;
 } catch (err) {
   highlightFn = text => {
     if (!cardinalRecommended) {
