@@ -10,8 +10,55 @@ class BinaryRow {
     this.columns = columns || [];
   }
 
-  toPacket() {
-    throw new Error('Not implemented');
+  static toPacket(columns, encoding) {
+    // throw new Error('Not implemented');
+    const sequenceId = 0; // TODO remove, this is calculated now in connecton
+    let length = 0;
+    columns.forEach(val => {
+      if (val === null || typeof val === 'undefined') {
+        ++length;
+        return;
+      }
+      length += Packet.lengthCodedStringLength(val.toString(10), encoding);
+    });
+
+    length = length + 2;
+
+    const buffer = Buffer.allocUnsafe(length + 4);
+    const packet = new Packet(sequenceId, buffer, 0, length + 4);
+    packet.offset = 4;
+
+    packet.writeInt8(0);
+
+    let bitmap = 0;
+    let bitValue = 1;
+    columns.forEach(parameter => {
+      if (parameter.type === Types.NULL) {
+        bitmap += bitValue;
+      }
+      bitValue *= 2;
+      if (bitValue === 256) {
+        packet.writeInt8(bitmap);
+        bitmap = 0;
+        bitValue = 1;
+      }
+    });
+    if (bitValue !== 1) {
+      packet.writeInt8(bitmap);
+    }
+
+    columns.forEach(val => {
+      if (val === null) {
+        packet.writeNull();
+        return;
+      }
+      if (typeof val === 'undefined') {
+        packet.writeInt8(0);
+        return;
+      }
+      packet.writeLengthCodedString(val.toString(10), encoding);
+    });
+    return packet;
   }
 
   // TODO: complete list of types...
