@@ -80,12 +80,35 @@ function readCodeFor(field, config, options, fieldNum) {
 
 function compile(fields, options, config) {
   const parserFn = genFunc();
-  let i = 0;
   const nullBitmapLength = Math.floor((fields.length + 7 + 2) / 8);
 
-  /* eslint-disable no-trailing-spaces */
-  /* eslint-disable no-spaced-func */
-  /* eslint-disable no-unexpected-multiline */
+  function wrap(field, packet) {
+    return {
+      type: typeNames[field.columnType],
+      length: field.columnLength,
+      db: field.schema,
+      table: field.table,
+      name: field.name,
+      string: function (encoding = field.encoding) {
+        if (field.columnType === Types.JSON && encoding === field.encoding) {
+          // Since for JSON columns mysql always returns charset 63 (BINARY),
+          // we have to handle it according to JSON specs and use "utf8",
+          // see https://github.com/sidorares/node-mysql2/issues/1661
+          console.warn(
+            `typeCast: JSON column "${field.name}" is interpreted as BINARY by default, recommended to manually set utf8 encoding: \`field.string("utf8")\``,
+          );
+        }
+
+        return packet.readLengthCodedString(encoding);
+      },
+      buffer: function () {
+        return packet.readLengthCodedBuffer();
+      },
+      geometry: function () {
+        return packet.parseGeometryValue();
+      },
+    };
+  }
 
   parserFn('(function(){');
   parserFn('return class BinaryRow {');
@@ -96,20 +119,15 @@ function compile(fields, options, config) {
   if (options.rowsAsArray) {
     parserFn(`const result = new Array(${fields.length});`);
   } else {
-    parserFn("const result = {};");
+    parserFn('const result = {};');
   }
 
-  const resultTables = {};
-  let resultTablesArray = [];
-
-  if (options.nestTables === true) {
-    for (i = 0; i < fields.length; i++) {
-      resultTables[fields[i].table] = 1;
-    }
-    resultTablesArray = Object.keys(resultTables);
-    for (i = 0; i < resultTablesArray.length; i++) {
-      parserFn(`result[${helpers.srcEscape(resultTablesArray[i])}] = {};`);
-    }
+  // Global typeCast
+  if (
+    typeof config.typeCast === 'function' &&
+    typeof options.typeCast !== 'function'
+  ) {
+    options.typeCast = config.typeCast;
   }
   
   // If we're processing dots, we need to pre-declare all objects that are used
@@ -147,7 +165,7 @@ function compile(fields, options, config) {
   }
 
   parserFn('packet.readInt8();'); // status byte
-  for (i = 0; i < nullBitmapLength; ++i) {
+  for (let i = 0; i < nullBitmapLength; ++i) {
     parserFn(`const nullBitmaskByte${i} = packet.readInt8();`);
   }
 
@@ -157,17 +175,17 @@ function compile(fields, options, config) {
   let fieldName = '';
   let tableName = '';
 
-  for (i = 0; i < fields.length; i++) {
+  for (let i = 0; i < fields.length; i++) {
     fieldName = helpers.srcEscape(fields[i].name);
     parserFn(`// ${fieldName}: ${typeNames[fields[i].columnType]}`);
 
     if (typeof options.nestTables === 'string') {
-      tableName = helpers.srcEscape(fields[i].table);
       lvalue = `result[${helpers.srcEscape(
-        fields[i].table + options.nestTables + fields[i].name
+        fields[i].table + options.nestTables + fields[i].name,
       )}]`;
     } else if (options.nestTables === true) {
       tableName = helpers.srcEscape(fields[i].table);
+      parserFn(`if (!result[${tableName}]) result[${tableName}] = {};`);
       lvalue = `result[${tableName}][${fieldName}]`;
     } else if (options.rowsAsArray) {
       lvalue = `result[${i.toString(10)}]`;
@@ -178,23 +196,29 @@ function compile(fields, options, config) {
       });
       lvalue = `this[${fieldNames.join('][')}]`;
     } else {
-      lvalue = `result[${helpers.srcEscape(fields[i].name)}]`;
+      lvalue = `result[${fieldName}]`;
     }
 
-    // TODO: this used to be an optimisation ( if column marked as NOT_NULL don't include code to check null
-    // bitmap at all, but it seems that we can't rely on this flag, see #178
-    // TODO: benchmark performance difference
-    //
-    // if (fields[i].flags & FieldFlags.NOT_NULL) { // don't need to check null bitmap if field can't be null.
-    //  result.push(lvalue + ' = ' + readCodeFor(fields[i], config));
-    // } else if (fields[i].columnType == Types.NULL) {
-    //  result.push(lvalue + ' = null;');
-    // } else {
-    parserFn(`if (nullBitmaskByte${nullByteIndex} & ${currentFieldNullBit})`);
-    parserFn(`${lvalue} = null;`);
-    parserFn('else');
-    parserFn(`${lvalue} = ${readCodeFor(fields[i], config, options, i)}`);
-    // }
+    if (options.typeCast === false) {
+      parserFn(`${lvalue} = packet.readLengthCodedBuffer();`);
+    } else {
+      const fieldWrapperVar = `fieldWrapper${i}`;
+      parserFn(`const ${fieldWrapperVar} = wrap(fields[${i}], packet);`);
+      const readCode = readCodeFor(fields[i], config, options, i);
+
+      parserFn(`if (nullBitmaskByte${nullByteIndex} & ${currentFieldNullBit})`);
+      parserFn(`${lvalue} = null;`);
+      parserFn('else {');
+      if (typeof options.typeCast === 'function') {
+        parserFn(
+          `${lvalue} = options.typeCast(${fieldWrapperVar}, function() { return ${readCode} });`,
+        );
+      } else {
+        parserFn(`${lvalue} = ${readCode};`);
+      }
+      parserFn('}');
+    }
+
     currentFieldNullBit *= 2;
     if (currentFieldNullBit === 0x100) {
       currentFieldNullBit = 1;
@@ -206,17 +230,13 @@ function compile(fields, options, config) {
   parserFn('}');
   parserFn('};')('})()');
 
-  /* eslint-enable no-trailing-spaces */
-  /* eslint-enable no-spaced-func */
-  /* eslint-enable no-unexpected-multiline */
-
   if (config.debug) {
     helpers.printDebugWithCode(
       'Compiled binary protocol row parser',
-      parserFn.toString()
+      parserFn.toString(),
     );
   }
-  return parserFn.toFunction();
+  return parserFn.toFunction({ wrap });
 }
 
 function getBinaryParser(fields, options, config) {
