@@ -27,6 +27,10 @@ class Pool extends EventEmitter {
     this._freeConnections = new Queue();
     this._connectionQueue = new Queue();
     this._closed = false;
+    if (this.config.maxIdle < this.config.connectionLimit) {
+      // create idle connection timeout automatically release job
+      this._removeIdleTimeoutConnections();
+    }
   }
 
   promise(promiseImpl) {
@@ -40,7 +44,7 @@ class Pool extends EventEmitter {
     }
     let connection;
     if (this._freeConnections.length > 0) {
-      connection = this._freeConnections.shift();
+      connection = this._freeConnections.pop();
       this.emit('acquire', connection);
       return process.nextTick(() => cb(null, connection));
     }
@@ -96,6 +100,7 @@ class Pool extends EventEmitter {
 
   end(cb) {
     this._closed = true;
+    clearTimeout(this._removeIdleTimeoutConnectionsTimer);
     if (typeof cb !== 'function') {
       cb = function(err) {
         if (err) {
@@ -174,7 +179,7 @@ class Pool extends EventEmitter {
         });
       } catch (e) {
         conn.release();
-        throw e;
+        return cb(e);
       }
     });
   }
@@ -185,6 +190,26 @@ class Pool extends EventEmitter {
     // Remove connection from free connections
     spliceConnection(this._freeConnections, connection);
     this.releaseConnection(connection);
+  }
+
+  _removeIdleTimeoutConnections() {
+    if (this._removeIdleTimeoutConnectionsTimer) {
+      clearTimeout(this._removeIdleTimeoutConnectionsTimer);
+    }
+
+    this._removeIdleTimeoutConnectionsTimer = setTimeout(() => {
+      try {
+        while (
+          this._freeConnections.length > this.config.maxIdle &&
+          Date.now() - this._freeConnections.get(0).lastActiveTime >
+            this.config.idleTimeout
+        ) {
+          this._freeConnections.get(0).destroy();
+        }
+      } finally {
+        this._removeIdleTimeoutConnections();
+      }
+    }, 1000);
   }
 
   format(sql, values) {
