@@ -12,7 +12,7 @@ const connection = common.createConnection();
 
 // Test that we ignore a ECONNRESET error if the connection
 // is already closing, we close and then emit the error
-connection.query(`select 1`, (err, rows) => {
+connection.query(`select 1 as "1"`, (err, rows) => {
   assert.equal(rows[0]['1'], 1);
   connection.close();
   connection.stream.emit('error', error);
