@@ -5,6 +5,12 @@
 
 'use strict';
 
+// different error codes for PS, disabling for now
+if (`${process.env.MYSQL_CONNECTION_URL}`.includes('pscale_pw_')) {
+  console.log('skipping test for planetscale');
+  process.exit(0);
+}
+
 const common = require('../../common');
 const connection = common.createConnection();
 const assert = require('assert-diff');
@@ -52,7 +58,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 63,
-    columnType: 8,
+    encoding: 'binary',
+    type: 8,
     flags: 161,
     decimals: 0
   },
@@ -64,7 +71,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 224,
-    columnType: 253,
+    encoding: 'utf8',
+    type: 253,
     flags: 1,
     decimals: 31
   },
@@ -76,7 +84,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 224,
-    columnType: 253,
+    encoding: 'utf8',
+    type: 253,
     flags: 0,
     decimals: 31
   },
@@ -88,7 +97,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 224,
-    columnType: 250,
+    encoding: 'utf8',
+    type: 250,
     flags: 0,
     decimals: 31
   },
@@ -100,7 +110,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 224,
-    columnType: 253,
+    encoding: 'utf8',
+    type: 253,
     flags: 0,
     decimals: 31
   },
@@ -112,7 +123,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 224,
-    columnType: 253,
+    encoding: 'utf8',
+    type: 253,
     flags: 0,
     decimals: 31
   },
@@ -124,7 +136,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 224,
-    columnType: 253,
+    encoding: 'utf8',
+    type: 253,
     flags: 0,
     decimals: 31
   },
@@ -136,7 +149,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 224,
-    columnType: 253,
+    encoding: 'utf8',
+    type: 253,
     flags: 0,
     decimals: 31
   },
@@ -148,7 +162,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 224,
-    columnType: 253,
+    encoding: 'utf8',
+    type: 253,
     flags: 0,
     decimals: 31
   },
@@ -160,7 +175,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 63,
-    columnType: 8,
+    encoding: 'binary',
+    type: 8,
     flags: 160,
     decimals: 0
   },
@@ -172,7 +188,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 63,
-    columnType: 5,
+    encoding: 'binary',
+    type: 5,
     flags: 128,
     decimals: 2
   },
@@ -184,7 +201,8 @@ const expectedFields = [
     table: '',
     orgTable: '',
     characterSet: 224,
-    columnType: 253,
+    encoding: 'utf8',
+    type: 253,
     flags: 1,
     decimals: 31
   }
