@@ -4,16 +4,17 @@ const common = require('../../common');
 const connection = common.createConnection();
 const assert = require('assert');
 
-let rows;
-connection.query('SELECT ""', (err, _rows) => {
+let rows, fields;
+connection.query('SELECT ""', (err, _rows, _fields) => {
   if (err) {
     throw err;
   }
 
   rows = _rows;
+  fields = _fields;
   connection.end();
 });
 
 process.on('exit', () => {
-  assert.deepEqual(rows, [{ '': '' }]);
+  assert.deepEqual(rows, [{ [fields[0].name]: '' }]);
 });
