@@ -5,18 +5,15 @@ const connection = common.createConnection();
 const assert = require('assert');
 
 let rows = undefined;
-let fields = undefined;
 const multibyteText = '本日は晴天なり';
-connection.query(`SELECT '${multibyteText}'`, (err, _rows, _fields) => {
+connection.query(`SELECT '${multibyteText}' as result`, (err, _rows) => {
   if (err) {
     throw err;
   }
   rows = _rows;
-  fields = _fields;
   connection.end();
 });
 
 process.on('exit', () => {
-  assert.equal(rows[0][multibyteText], multibyteText);
-  assert.equal(fields[0].name, multibyteText);
+  assert.equal(rows[0].result, multibyteText);
 });
