@@ -1,24 +1,18 @@
 'use strict';
 
-const code = `
-'use strict';
-
 const config = require('../../common.js').config;
 const assert = require('assert');
-const isAsyncSupported = require('is-async-supported');
 const ErrorStackParser = require('error-stack-parser');
 
-let skipTest = false;
-if (!isAsyncSupported) {
-  console.log('no async/await support, skipping test');
-  skipTest = true;
-  process.exit(0);
-}
+const createConnection = async function(args) {
+  const connect = require('../../../promise.js').createConnection;
+  if (!args && process.env.MYSQL_CONNECTION_URL) {
+    return connect({ uri: process.env.MYSQL_CONNECTION_URL });
+  }
+  return connect({ ...config, ...args });
+};
 
-const createConnection = require('../../../promise.js').createConnection;
-const createPool = require('../../../promise.js').createPool;
-
-function test() {
+async function test() {
   // TODO check this is actially required. This meant as a help for pre async/await node
   // to load entire file and do isAsyncSupported check instead of failing with syntax error
 
@@ -27,22 +21,12 @@ function test() {
   // TODO: investigate why connection is still open after ENETUNREACH
   async function test1() {
     e1 = new Error();
-    const conn = await createConnection({ host: '0.42.42.42' });
-    let [rows, fields] = conn.query('select 1 + 1');
-    await Promise.all([conn.query('select 1+1'), conn.query('syntax error')]);
+    // expected not to connect
+    await createConnection({ host: '127.0.0.1', port: 33066 });
   }
 
-  /*
-  test1().catch(err => {
-    const stack = ErrorStackParser.parse(err);
-    const stackExpected = ErrorStackParser.parse(e1);
-    assert(stack[1].getLineNumber() === stackExpected[0].getLineNumber() + 1);
-  });
-  */
-
   async function test2() {
-    const conn = await createConnection(config);
-    let [rows, fields] = await conn.query('select 1 + 1');
+    const conn = await createConnection();
     try {
       e2 = new Error();
       await Promise.all([conn.query('select 1+1'), conn.query('syntax error')]);
@@ -54,30 +38,12 @@ function test() {
     }
   }
 
-  test2();
+  test1().catch(err => {
+    const stack = ErrorStackParser.parse(err);
+    const stackExpected = ErrorStackParser.parse(e1);
+    assert(stack[2].getLineNumber() === stackExpected[0].getLineNumber() + 2);
+    test2();
+  });
 }
 
 test();
-`;
-
-process.on('unhandledRejection', err => {
-  console.log(err.stack);
-});
-
-const vm = require('vm');
-
-try {
-  vm.runInNewContext(
-    code,
-    {
-      require: require
-    },
-    {
-      fileName: __filename,
-      lineOffset: 1
-    }
-  );
-} catch (err) {
-  // ignore sync errors (must be syntax - async/await not supported)
-  console.log(err);
-}
