@@ -37,7 +37,6 @@ packet = ColumnDefinition.toPacket(
     orgName: 'on_погоди',
     table: 't_погоди',
     orgTable: 'ot_погоди',
-
     characterSet: 0x21,
     columnLength: 500,
     flags: 32896,
@@ -64,6 +63,8 @@ const inputColDef = {
   columnLength: 0x1c,
   flags: 0,
   columnType: 0xfd,
+  type: 0xfd,
+  encoding: 'latin1',
   decimals: 0x1f
 }
 packet = ColumnDefinition.toPacket(inputColDef, sequenceId);
@@ -74,6 +75,9 @@ assert.equal(
 
 packet.offset = 4;
 const colDef = new ColumnDefinition(packet, "utf8");
-const inspect = colDef.inspect();
+// inspect omits the "colulumnType" property because type is an alias for it
+// but ColumnDefinition.toPacket reads type from "columnType"
+// TODO: think how to make this more consistent
+const inspect = { columnType: 253, ...colDef.inspect() };
 assert.deepEqual(inspect, inputColDef);
 assert.equal(colDef.db, inputColDef.schema);
