@@ -1,22 +1,27 @@
-
 declare interface FieldPacket {
-    constructor: {
-        name: 'FieldPacket'
-    };
-    catalog: string;
-    charsetNr: number;
-    db: string;
-    decimals: number;
-    default: any;
-    flags: number;
-    length: number;
-    name: string;
-    orgName: string;
-    orgTable: string;
-    protocol41: boolean;
-    table: string;
-    type: number;
-    zerofill: boolean;
+  constructor: {
+    name: 'FieldPacket';
+  };
+  catalog: string;
+  charsetNr?: number;
+  db?: string;
+  schema?: string;
+  characterSet?: number;
+  decimals: number;
+  default?: any;
+  flags: number | string[];
+  length?: number;
+  name: string;
+  orgName: string;
+  orgTable: string;
+  protocol41?: boolean;
+  table: string;
+  type?: number;
+  columnType?: number
+  zerofill?: boolean;
+  typeName?: string;
+  encoding?: string;
+  columnLength?: number;
 }
 
-export = FieldPacket;
+export { FieldPacket };
