@@ -1,10 +1,9 @@
-
 declare interface RowDataPacket {
-    constructor: {
-        name: 'RowDataPacket'
-    };
-    [column: string]: any;
-    [column: number]: any;
+  constructor: {
+    name: 'RowDataPacket';
+  };
+  [column: string]: any;
+  [column: number]: any;
 }
 
-export = RowDataPacket;
+export { RowDataPacket };
