@@ -1,5 +1,5 @@
+import { EventEmitter } from 'events';
 
-import {EventEmitter} from 'events';
+declare class Sequence extends EventEmitter {}
 
-declare class Sequence extends EventEmitter { }
-export = Sequence;
+export { Sequence };
